segmentation/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
_py-xgboost-mutex=2.0=cpu_0
absl-py=0.10.0=py37_0
attrs=20.2.0=py_0
backcall=0.2.0=py_0
beautifulsoup4=4.9.3=pyhb0f4dca_0
blas=1.0=mkl
bleach=3.2.1=py_0
brotlipy=0.7.0=py37h7b6447c_1000
bzip2=1.0.8=h7b6447c_0
c-ares=1.16.1=h7b6447c_0
ca-certificates=2022.12.7=ha878542_0
captum=0.3.1=0
certifi=2022.12.7=pyhd8ed1ab_0
cffi=1.14.3=py37he30daa8_0
chardet=3.0.4=py37_1003
cloudpickle=1.6.0=py_0
colour=0.1.5=py_0
cryptography=3.1.1=py37h1ba5d50_0
cudatoolkit=11.3.1=h2bc3f7f_2
cudatoolkit-dev=11.6.0=h72bdee0_5
cupy=10.3.0=py37h17e6d27_0
cycler=0.10.0=py37_0
cytoolz=0.11.0=py37h7b6447c_0
dask-core=2.30.0=py_0
dbus=1.13.18=hb2f20db_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
easydict=1.9=py_0
einops=0.6.0=pyhd8ed1ab_0
entrypoints=0.3=py37_0
expat=2.2.10=he6710b0_2
fastrlock=0.8=py37hd23a5d3_1
ffmpeg=4.3=hf484d3e_0
filelock=3.0.12=py_0
fontconfig=2.13.0=h9420a91_0
freetype=2.10.4=h5ab3b9f_0
gdown=4.2.0=pyhd8ed1ab_0
glib=2.56.2=hd408876_0
gmp=6.1.2=hb3b607b_0
gnutls=3.6.5=h71b1129_1002
grpcio=1.31.0=py37hf8bcb03_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
h5py=2.10.0=py37hd6299e0_1
hdf5=1.10.6=hb1b8bf9_0
icu=58.2=he6710b0_3
idna=2.10=py_0
imageio=2.9.0=py_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
intel-openmp=2020.2=254
iprogress=0.4=py_0
ipykernel=5.3.4=py37h5ca1d4c_0
ipython=7.18.1=py37h5ca1d4c_0
ipython_genutils=0.2.0=py37_0
ipywidgets=7.6.5=pyhd8ed1ab_0
jedi=0.18.1=py37h89c1867_0
jinja2=2.11.2=py_0
joblib=0.17.0=py_0
jpeg=9b=habf39ab_1
jsonschema=3.2.0=py_2
jupyter_client=6.1.7=py_0
jupyter_core=4.6.3=py37_0
jupyterlab_widgets=1.0.2=pyhd8ed1ab_0
kiwisolver=1.2.0=py37hfd86e86_0
kornia=0.6.2=pyhd8ed1ab_0
lame=3.100=h7b6447c_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1d223b6_11
libgfortran-ng=7.3.0=hdf63c60_0
libgomp=11.2.0=h1d223b6_11
libiconv=1.15=h63c8f33_5
libllvm10=10.0.1=hbcb73fb_5
libpng=1.6.37=hbc83047_0
libprotobuf=3.13.0.1=hd408876_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=he4da1e4_11
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libxcb=1.14=h7b6447c_0
libxgboost=1.3.3=h9c3ff4c_2
libxml2=2.9.10=hb55368b_3
lime=0.2.0.1=pyhd8ed1ab_1
llvmlite=0.34.0=py37h269e1b5_4
lz4-c=1.9.2=heb0550a_3
markdown=3.3.2=py37_0
markupsafe=1.1.1=py37h14c3975_1
matplotlib=3.3.1=0
matplotlib-base=3.3.1=py37h817c723_0
mistune=0.8.4=py37h14c3975_1001
mkl=2019.4=243
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.2.0=py37h23d657b_0
mkl_random=1.0.4=py37hd81dba3_0
nbconvert=5.6.1=py37_1
nbformat=5.0.8=py_0
ncurses=6.2=he6710b0_1
nettle=3.4.1=hbb512f6_0
networkx=2.5=py_0
ninja=1.10.1=py37hfd86e86_0
notebook=5.7.10=py37hc8dfbb8_0
numba=0.51.2=py37h04863e7_1
numpy=1.19.1=py37hbc911f0_0
numpy-base=1.19.1=py37hfa32c7d_0
olefile=0.46=py37_0
open-fonts=0.7.0=1
openh264=2.1.0=hd408876_0
openssl=1.1.1l=h7f98852_0
packaging=20.4=py_0
pandas=1.1.3=py37he6710b0_0
pandoc=2.11=hb0f4dca_0
pandocfilters=1.4.2=py37_1
parso=0.8.0=py_0
pcre=8.44=he6710b0_0
pexpect=4.8.0=py37_1
pickleshare=0.7.5=py37_1001
pillow=8.0.0=py37h9a89aac_0
pip=21.1.3=py37h06a4308_0
prometheus_client=0.8.0=py_0
prompt-toolkit=3.0.8=py_0
protobuf=3.13.0.1=py37he6710b0_1
ptyprocess=0.6.0=py37_0
py-xgboost=1.3.3=py37h89c1867_2
pycparser=2.20=py_2
pygments=2.7.1=py_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=py_0
pyqt=5.9.2=py37h22d08a2_1
pyrsistent=0.17.3=py37h7b6447c_0
pysocks=1.7.1=py37_1
python=3.7.9=h7579374_0
python-dateutil=2.8.1=py_0
python_abi=3.7=2_cp37m
pytorch=1.11.0=py3.7_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pytz=2020.1=py_0
pywavelets=1.1.1=py37h7b6447c_2
pyyaml=5.3.1=py37h7b6447c_1
pyzmq=19.0.2=py37he6710b0_1
qt=5.9.7=h5867ecd_1
rclone=1.57.0=h519d9b9_0
readline=8.0=h7b6447c_0
requests=2.24.0=py_0
scikit-image=0.17.2=py37hdf5156a_0
scikit-learn=0.23.2=py37h0573a6f_0
scipy=1.6.1=py37h91f5cce_0
seaborn=0.11.0=py_0
send2trash=1.5.0=py37_0
setuptools=50.3.0=py37hb0f4dca_1
shap=0.37.0=py37h10a2094_0
sip=4.19.24=py37he6710b0_0
six=1.15.0=py_0
slicer=0.0.7=pyhd8ed1ab_0
soupsieve=2.0.1=py_0
sqlite=3.33.0=h62c20be_0
tbb=2020.3=hfd86e86_0
tensorboard=1.15.0=py37_0
tensorboardx=2.4=pyhd8ed1ab_0
terminado=0.9.1=py37_0
testpath=0.4.4=py_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tifffile=2020.10.1=py37hdd07704_2
tk=8.6.10=hbc83047_0
toolz=0.11.1=py_0
torchfile=0.1.0=py_0
torchvision=0.12.0=py37_cu113
tornado=6.0.4=py37h7b6447c_1
tqdm=4.56.2=pyhd8ed1ab_0
traitlets=5.0.5=py_0
typing_extensions=3.7.4.3=py_0
unzip=6.0=h7f98852_2
urllib3=1.25.11=py_0
visdom=0.1.8.9=0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py37_1
websocket-client=0.57.0=py37_2
werkzeug=1.0.1=py_0
wheel=0.35.1=py_0
widgetsnbextension=3.5.1=py37_0
xgboost=1.3.3=py37h89c1867_2
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zeromq=4.3.3=he6710b0_3
zip=3.0=h7f98852_1
zipp=3.3.1=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.4=h0b5b093_3
